aws-auth0-auth/requirements.txt
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.1
ecdsa==0.13
enum34==1.1.6
future==0.16.0
http==0.2
httpclient==0.0.2
idna==2.6
ipaddress==1.0.18
logging==0.4.9.6
pycparser==2.18
pycrypto==2.6.1
python-jose==1.4.0
requests==2.18.4
six==1.11.0
urllib3==1.22
